requirements.txt
django
djangorestframework
django-cors-headers
nltk
beautifulsoup4
drfdocs
pypdf2